+++ packages/at_chops/lib/src/algorithm/encryption/ml_kem_768_ffi.dart
@@ -3,9 +3,9 @@ import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:at_chops/src/algorithm/at_algorithm.dart';
-import 'package:at_chops/src/algorithm/encryption/ml_kem_768_validation.dart';
 import 'package:at_chops/src/algorithm/ffi/openssl_ffi_bindings.dart';
-import 'package:at_chops/src/algorithm/pq_validation.dart';
+import 'package:at_chops/src/algorithm/spec/ml_kem_768_spec.dart';
+import 'package:at_chops/src/algorithm/spec/output_length.dart';
 import 'package:ffi/ffi.dart';
 
 /// ML-KEM-768 (FIPS 203) KEM backed by OpenSSL 3 via Dart FFI.
