+++ packages/at_chops/lib/src/algorithm/encryption/x_wing_pure_dart.dart
@@ -4,10 +4,10 @@ import 'dart:typed_data';
 
 import 'package:at_chops/src/algorithm/at_algorithm.dart';
 import 'package:at_chops/src/algorithm/encryption/ml_kem_768_pure_dart.dart';
-import 'package:at_chops/src/algorithm/encryption/ml_kem_768_validation.dart';
 import 'package:at_chops/src/algorithm/encryption/x25519_pure_dart_algo.dart';
-import 'package:at_chops/src/algorithm/encryption/x_wing_sizes.dart';
-import 'package:at_chops/src/algorithm/pq_validation.dart';
+import 'package:at_chops/src/algorithm/spec/ml_kem_768_spec.dart';
+import 'package:at_chops/src/algorithm/spec/output_length.dart';
+import 'package:at_chops/src/algorithm/spec/x_wing_spec.dart';
 import 'package:cryptography/cryptography.dart' as crypto;
 import 'package:meta/meta.dart';
 import 'package:pointycastle/digests/sha3.dart';
@@ -183,9 +183,9 @@ final class XWingPureDartAlgo implements AtKemAlgorithm {
   /// exactly 32 bytes; checking it here covers both call sites in one place.
   Uint8List _combine(
       Uint8List ssM, Uint8List ssX, Uint8List ctX, Uint8List pkX) {
-    checkOutputLength(ssM.length, 32,
+    checkOutputLength(ssM.length, MlKem768Sizes.sharedSecretBytes,
         operation: 'X-Wing', label: 'ML-KEM-768 shared secret component');
-    checkOutputLength(ssX.length, 32,
+    checkOutputLength(ssX.length, XWingSizes.x25519ComponentLength,
         operation: 'X-Wing', label: 'X25519 shared secret component');
     final Uint8List input =
         Uint8List(ssM.length + ssX.length + ctX.length + pkX.length + 6)
