+++ packages/at_chops/lib/src/algorithm/signing/ml_dsa_65_pure_dart.dart
@@ -3,7 +3,8 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:at_chops/src/algorithm/at_algorithm.dart';
-import 'package:at_chops/src/algorithm/pq_validation.dart';
+import 'package:at_chops/src/algorithm/spec/ml_dsa_65_spec.dart';
+import 'package:at_chops/src/algorithm/spec/output_length.dart';
 import 'package:at_commons/at_commons.dart';
 import 'package:pqcrypto/pqcrypto.dart';
