+++ packages/at_chops/lib/src/key/impl/ml_dsa_65_key_pair.dart
@@ -1,11 +1,11 @@
 import 'dart:convert';
 import 'dart:typed_data';
 
+import 'package:at_chops/src/algorithm/spec/ml_dsa_65_spec.dart';
 import 'package:at_chops/src/key/keys.dart';
 import 'package:at_commons/at_commons.dart';
 
 import '../../algorithm/signing/ml_dsa_65_pure_dart.dart';
-import '../../algorithm/pq_validation.dart';
 
 /// ML-DSA-65 (FIPS 204) signing key pair.
 ///
