+++ packages/at_chops/test/ml_dsa_65_algo_test.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:at_chops/at_chops.dart';
-import 'package:at_chops/src/algorithm/pq_validation.dart';
+import 'package:at_chops/src/algorithm/spec/ml_dsa_65_spec.dart';
 import 'package:at_commons/at_commons.dart';
 import 'package:pqcrypto/pqcrypto.dart';
 import 'package:test/test.dart';
