+++ packages/at_chops/test/ml_kem_768_ffi_test.dart
@@ -6,7 +6,7 @@ import 'dart:ffi';
 import 'dart:typed_data';
 
 import 'package:at_chops/at_chops_ffi.dart';
-import 'package:at_chops/src/algorithm/encryption/ml_kem_768_validation.dart';
+import 'package:at_chops/src/algorithm/spec/ml_kem_768_spec.dart';
 import 'package:test/test.dart';
 
 void main() {
